requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
pydantic
openpyxl
